renamed models file to match singular campaign finance naming convention
